requirements.txt
flask>=3.0.0
feedparser>=6.0.0
requests>=2.31.0
apscheduler>=3.10.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
trafilatura>=1.12.0
